requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
wandb>=0.15.0

# Install separately:
# pip install git+https://github.com/openai/CLIP.git
# pip install git+https://github.com/facebookresearch/segment-anything.git
# Follow Grounding DINO installation instructions
